pyproject.toml
[project]
name = "codename-mindmosaic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.6",
    "marvin>=2.3.8",
    "python-dotenv>=1.0.1",
    "sqlmodel>=0.0.22",
    "streamlit>=1.41.1",
    "uvicorn>=0.34.0",
]

[tool.uv]
dev-dependencies = [
    "flake8>=7.1.1",
    "pytest>=8.3.3",
    "pre-commit>=4.0.1",
    "ruff>=0.7.2",
]
